Revert " applied changes requested by tropicaaal"
This reverts commit f692015defdcb5c538de51564d9f43f1898f1aee.

# Conflicts:
#	Minecraft.World/Network/Packets/MoveEntityPacket.cpp

+++ Minecraft.World/Network/Packets/SetEntityMotionPacket.h
@@ -7,7 +7,7 @@ class SetEntityMotionPacket : public Packet, public std::enable_shared_from_this
 {
 public:
 	int id;
-	int16_t xa, ya, za;
+    int xa, ya, za;
 	bool useBytes;		// 4J added
 
 private:
